requirements.txt
torch torchvision torchaudio
transformers
sounddevice
numpy
soundfile
pydub
google-generativeai
ffmpeg-python
playsound
scipy
openai
fastapi
uvicon
uvicon-standard
